+++ CHANGELOG.md
@@ -39,3 +39,4 @@
 - Significantly increased viable spell depth, making it effectively™ infinite. (@StellarWitch7)
 - Impulse cost scaling now accounts for multiple casts within the same tick. (@StellarWitch7)
   - As a consequence, tricks may now internally attach arbitrary data to a spell which persists for the remaining executions in the current tick.
+- Added a trick to guarantee a spell's full execution within a single tick. (@StellarWitch7)